Rotatable cube with fixed colour mapping

+++ src/Material.cpp
@@ -15,7 +15,10 @@ Colour::Colour(uint8_t r, uint8_t g, uint8_t b, uint8_t a) {
 }
 
 uint32_t Colour::convertRGBA() const {
-    return (a << 24) | (r << 16) | (g << 8) | b;
+    return (static_cast<uint32_t>(a) << 24) |
+           (static_cast<uint32_t>(r) << 16) |
+           (static_cast<uint32_t>(g) << 8) |
+           static_cast<uint32_t>(b);
 }
 
 Material::Material() {
